Bug fix: Settings field used int instead of float

+++ src/components/MonostaticSensorSettings.tsx
@@ -237,7 +237,7 @@ export default function MonostaticSensorSettings() {
               type="number"
               value={sensor.receiver.rotation_time}
               onChange={(event) => {
-                const value = parseInt(event.target.value);
+                const value = parseFloat(event.target.value);
                 const newDetectableSensor = structuredClone(detectableSensor);
                 newDetectableSensor.sensor.receiver.rotation_time = value;
                 updateMonostaticSensor(newDetectableSensor);
